mononoke_api: list public bookmarks from the warm bookmarks cache
Summary:
For public bookmarks, we can avoid querying the database and instead serve
`list_bookmarks` from the warm bookmarks cache.  The listed bookmarks might
be slightly old, but only because derived data is still in progress, and so
listing the older bookmark value is a better choice.

The tests now need a way to make sure that the warm bookmark cache is
up-to-date, so add a `sync` method that waits for a complete cycle of the warm
bookmark cache update thread.

Reviewed By: StanislavGlebik

Differential Revision: D26693444

fbshipit-source-id: 7145964bdb8c22d98ab6a2bb8c5091c19addd03e