change hipster action in check_if_trusted
Summary:
It doesn't make a whole lot of sense to user the "tupperware" permission here.
Lets make a distinction in who is allowed to provision mononoke vs. who is allowed to forward identities to it.
I have created a separate action with the same identities allows as in the "tupperware" permission.

https://www.internalfb.com/intern/hipster/acls/view/?type=TIER&consumer=mononoke

Reviewed By: StanislavGlebik

Differential Revision: D26777489

fbshipit-source-id: d3999dc4da7a57ac721572610f65eba664e595e9