peers: allow remote peer ui to inherit help and edenapi configs
Summary:
We have configs that affect peer connections, like help.tlsauthhelp,
that are considered "repo-specific" configs now that they come from
dynamicconfig. Unfortunately repo-specific configs are removed from the ui when
copying it for use in a new remotepeer.

Let's add a few config sections to the allow list for what can go in a remote
peer ui. I have a task for making even repo-less commands load the standard
config, so in the future we can have these new peer objects use the standard
repo-less config, which will remove much of the need for maintaining this
allow-list.

Reviewed By: singhsrb

Differential Revision: D26784364

fbshipit-source-id: 30d9292e48b0f27ce7f4d4904ff6d5ff8dcaf069