mononoke/filenodes: group properly
Summary:
We use itertools group_by() to group filenodes. Unfortunately that only works
if they're sorted, which they're not. With a hundred shards, we easily end
up sending one query per filenode.

This fixes that to group the filenodes properly. This means we'll make 1 query
for 1000 filenodes, i.e. make 1000x fewer queries.

Differential Revision: D26821973

fbshipit-source-id: b83ccaff608f38adefe86c8f5018e4b525cbde11