mononoke/fileblob: don't silently ignore write errors
Summary:
Right now, if we get an error when writing to fileblob, we silently ignore it
under the assumption that it's an overwrite that failed. We shouldn't do this,
since it can cause to silently discard data.

For example, when using Mononoke with a tmpdir that isn't next to your blobs,
one might hit this error:

```
Os { code: 18, kind: Other, message: "Invalid cross-device link" }
```

Right now, we just silently ignore it and move on. With this diff, we'll
properly report that the write failed.

Reviewed By: farnz

Differential Revision: D26912990

fbshipit-source-id: 1b5a3c79c80f6b445e16e352e1c78aaec5e21850