segmented_changelog: update Manager to build OnDemandUpdate
Summary:
Our production setup has an OnDemandUpdateSegmentedChangelog that gets updated
in various ways. With a setup where the dag is reloaded completely from saves,
we need a factory for the OnDemandUpdateSegmentedChangelog.
SegmentedChangelogManager takes the role of being the factory for our
production Dags.

At some point we will remove the SegmentedChangelog implementation for Manager.

Reviewed By: krallin

Differential Revision: D26708173

fbshipit-source-id: b3d8ea612b317af374f2c0ce6d7c512e3b09b2d2