smartlog: add focused branch revset
Summary:
Add the `focusedbranch` revset.  This expands to the current draft branch, plus
any commits that are related to the branch via commit mutation and their
descendants.

This can be used as the revset for `hg smartlog` to give a more focused smartlog
that shows only the changesets relevant to the current stack.

Reviewed By: quark-zju

Differential Revision: D26852107

fbshipit-source-id: 9723935bf3bf900f8e109d4f507ae1db6627807e